refactor(dcc): split global-init parser out of dcc_func.c into dcc_global_init.c (§5)
Pure code movement: extract the ~1100-line file-scope object initializer
subsystem (parse_global_init_* / global_init_* / append_global_* and helpers)
from dcc_func.c into a new dcc_global_init.c translation unit. dcc_func.c
6215 -> 5108 lines.

The cluster was verified fully self-contained before moving: it references no
dcc_func.c file-static defined outside it, uses none of the file's other
file-scope static state, its own statics (incl. global_compound_literal_seq)
are used only within it, and all 11 non-static entry points were already
declared in dcc.h. No header/API changes were needed beyond documentation.

build-dcc.sh/.bat glob *.c (auto-included); CMakeLists.txt source list updated;
dcc.h module map updated. Behavior-preserving: 0 assembly diffs across 316
tests; suite 307 pass / 0 fail / 9 skip, diagnostics + performance unchanged.
